backend/requirements.txt
torch
torchsummary
numpy
torchvision
Pillow
flask